requirements.txt
Django=3.0,4.0
djangorestframework=3.12,4.0
psycopg2=2.8,3.0
django-cors-headers=3.7.0,4.0
